Remove clojurewerkz/support dependency.
This change is necessary due to the removal of crossovers from
lein-cljsbuild which is needed for clojurewerkz/support.core to work on
cljs. This also makes it easier for people to use it since they can't
forget the special build requirements (See #18).
Furthermore, validateur only uses one tiny function of the entire core
library which and pulls in an entire jar for that.
It also reduces the generated javascript code.

This patch reduces complexity and the library size as well as makes it
easier to use validateur.